refactor(): refactored utilities folder and files
